==> 1_Using_Chatbot_API/0_Starter_code.ipynb <==
# Download the necessary data for the RAG exercise
import pandas as pd

# Downloading and loading the mini blog dataset
url = "https://raw.githubusercontent.com/AlaFalaki/tutorial_notebooks/main/data/mini-llama-articles.csv"
mini_dataset = pd.read_csv(url)

# Displaying the first few entries (run in a Jupyter Notebook)
mini_dataset

mini_dataset.to_csv("mini-llama-articles.csv", index=False, sep="\t")
