==> 1_Using_Chatbot_API/1.5_Process_data_embeddings.ipynb <==
# Split the input text into chunks of specified size (1024)
def split_into_chunks(text, chunk_size=1024):
  chunks = []
  for i in range(0, len(text), chunk_size):
    chunks.append(text[i:i+chunk_size])

  return chunks

# Read Data
import csv

chunks = []

# Load the file as a CSV
with open("./1_Using_Chatbot_API/mini-llama-articles.csv", mode="r", encoding="utf-8") as file:
  csv_reader = csv.reader(file, delimiter="\t")

  for idx, row in enumerate( csv_reader ):
    if idx == 0: continue; # Skip header row
    chunks.extend(split_into_chunks(row[1]))

print("number of articles:", idx)
print("number of chunks:", len(chunks))

import pandas as pd

# Convert the list to a Pandas Dataframe
df = pd.DataFrame(chunks, columns=['chunk'])

print(df.keys())

# Generate data embeddings using the OpenAI text-embedding-3-small model

from openai import OpenAI

client = OpenAI()

# Defining a function that converts a text to embedding vector using OpenAI's Ada model.
def get_embedding(text):
  try:
    # Remove newlines
    text = text.replace("\n", " ")
    res = client.embeddings.create(input=[text], model="text-embedding-3-small")

    return res.data[0].embedding

  except:
        return None

from tqdm.notebook import tqdm
import numpy as np

# Generate embedding
print("Generating embeddings...")
embeddings = []
for index, row in tqdm(df.iterrows()):
  embeddings.append(get_embedding(row['chunk']))

# Add the "embedding" column to the dataframe
embeddings_values = pd.Series(embeddings)
df.insert(loc=1, column='embedding', value=embeddings_values)

from sklearn.metrics.pairwise import cosine_similarity

# The question we want to ask the model.
QUESTION = "How many parameters LLaMA2 model has?"
QUESTION_emb = get_embedding(QUESTION)

# The similarity between the questions and each part of the essay.
cosine_similarities = cosine_similarity([QUESTION_emb], df['embedding'].tolist())

print(cosine_similarities)

# Find the N highest scored chunks
import numpy as np

number_of_chunks_to_retrieve = 3

# Sort and find the index of N highest scored chunks
indices = np.argsort(cosine_similarities[0])[::-1][:number_of_chunks_to_retrieve]
print(indices)

# Look at the highest scored retrieved pieces of text
for idx, item in enumerate(df.chunk[indices]):
  print(f"> Chunk {idx+1}")
  print(item)
  print("----")

df.to_csv("mini-llama-articles-with-embeddings.csv", index=False, sep="\t", encoding="utf-8")
